requirements.txt
absl-py==2.0.0
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
executing==2.0.1
flatbuffers==23.5.26
fonttools==4.46.0
gast==0.5.4
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
idna==3.6
ipython==8.18.1
jedi==0.19.1
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.7
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.41
protobuf==4.23.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
threadpoolctl==3.2.0
traitlets==5.14.0
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1
